Loading the algorithm for multiplying large numbers
In this algorithm, very large numbers can be multiplied together.
In this project, which is written in C# and at the Windows Forms level, the user is able to give different numbers to the software and see the result of multiplying two input numbers displayed in a list. In addition, the software is able to save the input and output data in a json file.